Move FallbackExceptionHandler to root package
