+++ src/main/java/net/sf/jabref/JabRefMain.java
@@ -7,7 +7,6 @@ import javax.swing.SwingUtilities;
 
 import net.sf.jabref.cli.ArgumentProcessor;
 import net.sf.jabref.gui.remote.JabRefMessageHandler;
-import net.sf.jabref.gui.util.FallbackExceptionHandler;
 import net.sf.jabref.logic.CustomEntryTypesManager;
 import net.sf.jabref.logic.exporter.ExportFormat;
 import net.sf.jabref.logic.exporter.ExportFormats;
